fix: 🐛 move crowdfunding function getChainInfoByChainId,checkSupportNetwork to common
