+++ packages/web/src/pages/bounty/detail/components/Payment/applyBounty.tsx
@@ -2,12 +2,17 @@ import { UButton } from '@comunion/components'
 import { defineComponent, ref } from 'vue'
 import { ApplyDialog } from '../Dialog'
 import { APPLICANT_STATUS } from '@/constants'
+import { checkSupportNetwork } from '@/utils/wallet'
 
 export default defineComponent({
   props: {
     disabled: {
       type: Boolean,
       default: () => false
+    },
+    chainId: {
+      type: Number,
+      default: () => 0
     },
     applicantDepositMinAmount: {
       type: Number,
@@ -25,8 +30,11 @@ export default defineComponent({
     }
   },
   render() {
-    const applyBounty = () => {
-      this.applyBountyDialogVisible = !this.applyBountyDialogVisible
+    const applyBounty = async () => {
+      const isSupport = await checkSupportNetwork(this.chainId)
+      if (isSupport) {
+        this.applyBountyDialogVisible = !this.applyBountyDialogVisible
+      }
     }
     return (
       <>
