+++ packages/web/src/pages/crowdfunding/components/CrowdfundingInfo.tsx
@@ -3,10 +3,10 @@ import dayjs from 'dayjs'
 import { ethers } from 'ethers'
 import { defineComponent, PropType, ref, computed } from 'vue'
 import { CoinType } from '../[id]'
-import { getChainInfoByChainId } from '../utils'
 import { useErc20Contract } from '@/contracts'
 import { ServiceReturn } from '@/services'
 import { useWalletStore } from '@/stores'
+import { getChainInfoByChainId } from '@/utils/etherscan'
 
 export const CrowdfundingInfo = defineComponent({
   name: 'CrowdfundingInfo',
