+++ packages/web/src/pages/crowdfunding/components/Invest.tsx
@@ -4,11 +4,12 @@ import dayjs from 'dayjs'
 import { BigNumber, Contract, ethers } from 'ethers'
 import { defineComponent, computed, PropType, ref, onMounted } from 'vue'
 import { CoinType } from '../[id]'
-import { CrowdfundingStatus, getChainInfoByChainId } from '../utils'
+import { CrowdfundingStatus } from '../utils'
 import { useErc20Contract, useCrowdfundingContract } from '@/contracts'
 import { ServiceReturn, services } from '@/services'
 import { useUserStore, useWalletStore } from '@/stores'
 import { useContractStore } from '@/stores/contract'
+import { getChainInfoByChainId } from '@/utils/etherscan'
 import { formatToFloor } from '@/utils/numberFormat'
 import './invest.css'
