+++ packages/web/src/pages/crowdfunding/utils.ts
@@ -1,9 +1,3 @@
-import { allNetworks } from '@/constants'
-
-export const getChainInfoByChainId = (chainId: number) => {
-  return allNetworks.find(network => network.chainId === chainId)
-}
-
 export enum CrowdfundingStatus {
   UPCOMING = 1,
   LIVE = 2,
