+++ packages/web/src/utils/etherscan.ts
@@ -1,4 +1,5 @@
 import axios from 'axios'
+import { allNetworks } from '@/constants'
 
 const etherscanApiKey: string =
   (import.meta.env.VITE_ETHERSCAN_API_KEY as string) ?? 'M6CJEAZ7U6HSHAY1G2D84AK5C2S5U134JR'
@@ -7,6 +8,10 @@ let currentNetwork = 'Mainnet'
 
 export function setNetwork(network: 'Mainnet' | 'Goerli') {
   currentNetwork = network
+}
+
+export function getChainInfoByChainId(chainId: number) {
+  return allNetworks.find(network => network.chainId === chainId)
 }
 
 function getEndpintUrl() {
